Moved wray to its own directory to avoid recursive issue per #334; Fixed imports to reflect changes
