+++ isofit/radiative_transfer/sRTMnet.py
@@ -37,7 +37,7 @@ from isofit.configs.sections.radiative_transfer_config import (
 from isofit.core.common import VectorInterpolator, load_wavelen, resample_spectrum
 from isofit.core.sunposition import sunpos
 from isofit.radiative_transfer.six_s import SixSRT
-from isofit.utils.wrapped_ray import wray as ray
+from isofit.wrappers import ray
 
 from .look_up_tables import TabularRT
