+++ isofit/wrappers/__init__.py
@@ -1,0 +1,3 @@
+from .ray import Ray
+
+ray = Ray()
